fixed easyrsa install
wget command does not work and throws an error. "/tmp/EasyRSA-3.0.1-pivpn1/: Is a directory"
This is because the the output of wget is a directory, not a file. -O expects a file.
BN also, mv throws an error: "mv: cannot move `/tmp/EasyRSA-3.0.1-pivpn1' to `/etc/openvpn/easy-rsa': No such file or directory"
We can do a mkdir & wget & tar and move to do it in lesser steps, although less readable there are is no need for /tmp