Fixed status command (echo)
The status reported by `isRepo` fails to print a new line when the directory is not a git repo, because the function returns without printing the conclusion of the checks in one of the possible situations. This fixes this small visual bug. I don't know if the printed messages is the most suitable, but it now includes a new line.